requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.25.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.4
Django==3.2.2
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
